1.OrderService: implemented makeAnOrder method, with creation of an Order object 2.SearchView: Creditcard, list of passengers, meals, suitcases 3.Reduce flight seats (flighRepositoryImpl) when ordering a flight
1. Create a new Passenger and check if it does not exist.
2. Create a new order.
3. Decrease remaining flight seats in flights which in order.
4. Fixed all class containment in the project.
5. Added credit card validation.
6. Writing all order related data to the relevant files.

+++ src/model/objects/Airport.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class Airport {
 
     static int airportsCount;
-    private int id;
+    private long id;
     private String country, airportName;
     private List<String> terminals = new ArrayList<>();
 
@@ -26,7 +26,7 @@ public class Airport {
     public void setAirportName(String airportName) {
         this.airportName = airportName;
     }
-    public void setId(int id) { this.id = id; }
+    public void setId(long id) { this.id = id; }
 
     public String getCountry() {
         return this.country;
@@ -35,7 +35,7 @@ public class Airport {
         return this.airportName;
     }
     public List<String> getTerminals() { return this.terminals; }
-    public int getId() { return this.id; }
+    public long getId() { return this.id; }
 
 
     public boolean addTerminal(String terminal) {
